Enforce int for cdim

+++ make_test_snap.py
@@ -120,7 +120,7 @@ def make_ics(filepath, cdim, gdim, boxsize, doner_path):
 
         # Write the cells metadata
         cell_meta = cell_struct.create_group("Meta-data")
-        cell_meta.attrs["dimension"] = np.array([cdim, cdim, cdim])
+        cell_meta.attrs["dimension"] = np.array([cdim, cdim, cdim], dtype=int)
         cell_meta.attrs["size"] = np.array([cell_size, cell_size, cell_size])
 
         # Get the units from the doner snapshot
